[DellEMC]: Platform modules Python3 compliance and other changes (#5609)
- Make DellEMC platform modules Python3 compliant.
- Change return type of PSU Platform APIs in DellEMC Z9264, S5232 and Thermal Platform APIs in S5232 to 'float'.
- Remove multiple copies of pcisysfs.py.
- PEP8 style changes for utility scripts.
- Build and install Python3 version of sonic_platform package.
- Fix minor Platform API issues.

+++ platform/broadcom/sonic-platform-modules-dell/s6000/sonic_platform/psu.py
@@ -168,7 +168,7 @@ class Psu(PsuBase):
         status = False
         psu_status = self._get_cpld_register(self.psu_status_reg)
         if (psu_status != 'ERR'):
-            psu_status = (int(psu_status, 16) >> ((2 - self.index) * 4)) & 0xF
+            psu_status = (int(psu_status, 16) >> int((2 - self.index) * 4)) & 0xF
             if (~psu_status & 0b1000) and (~psu_status & 0b0100):
                 status = True
